Move DrainReclaimer and clean up local method to match others
